Move graph code into own file
dig.go is quite bloated and should be trimmed down a bit to
bare essentials, the rest can be grouped in logical files.
Starting with the graph felt right as I've scrolled over
that code at least a dozen times today.
